// src/domains/service-clients/course/proto/generated/course_service.ts
export interface Instructor {
  id: string;
  name: string;
  avatar?: string | undefined;
  email?: string | undefined;
}
// ...
export interface CourseData {
  id: string;
  title?: string | undefined;
  topics: string[];
  instructorId?: string | undefined;
  subTitle?: string | undefined;
  category?: string | undefined;
  subCategory?: string | undefined;
  language?: string | undefined;
  subtitleLanguage?: string | undefined;
  level?: string | undefined;
  durationValue?: string | undefined;
  durationUnit?: string | undefined;
  description?: string | undefined;
  learningOutcomes: string[];
  targetAudience: string[];
  requirements: string[];
  thumbnail?: string | undefined;
  trailer?: string | undefined;
  status?: string | undefined;
  slug?: string | undefined;
  rating?: number | undefined;
  numberOfRating?: number | undefined;
  enrollments?: number | undefined;
  sections: SectionData[];
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
  price?: number | undefined;
  discountPrice?: number | undefined;
  currency?: string | undefined;
  instructor?: Instructor | undefined;
}
// ...
export interface SectionData {
  id: string;
  courseId: string;
  title: string;
  lessons: LessonData[];
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
  description: string;
  isPublished: boolean;
  order: number;
  quiz: QuizData | undefined;
}
// ...
export interface LessonData {
  id: string;
  sectionId: string;
  isPreview: boolean;
  description: string;
  estimatedDuration: number;
  order: number;
  title: string;
  isPublished: boolean;
  contentType: string;
  contentUrl: string;
  metadata: ContentMetaData | undefined;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
}
// ...
export interface QuizData {
  id: string;
  courseId: string;
  sectionId: string;
  title: string;
  description?: string | undefined;
  timeLimit: number;
  passingScore: number;
  questions: Question[];
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
}
// ...
export const Instructor: MessageFns<Instructor> = {
  encode(message: Instructor, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.name !== "") {
      writer.uint32(18).string(message.name);
    }
    if (message.avatar !== undefined) {
      writer.uint32(26).string(message.avatar);
    }
    if (message.email !== undefined) {
      writer.uint32(34).string(message.email);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Instructor {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseInstructor();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.avatar = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.email = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Instructor {
    return {
      id: isSet(object.id) ? globalThis.String(object.id) : "",
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      avatar: isSet(object.avatar) ? globalThis.String(object.avatar) : undefined,
      email: isSet(object.email) ? globalThis.String(object.email) : undefined,
    };
  },

  toJSON(message: Instructor): unknown {
    const obj: any = {};
    if (message.id !== "") {
      obj.id = message.id;
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.avatar !== undefined) {
      obj.avatar = message.avatar;
    }
    if (message.email !== undefined) {
      obj.email = message.email;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Instructor>, I>>(base?: I): Instructor {
    return Instructor.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Instructor>, I>>(object: I): Instructor {
    const message = createBaseInstructor();
    message.id = object.id ?? "";
    message.name = object.name ?? "";
    message.avatar = object.avatar ?? undefined;
    message.email = object.email ?? undefined;
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}